Move frontend generator into client-cli
