+++ packages/client-cli/test/fixtures/frontend-openapi.json
