+++ packages/client-cli/test/fixtures/sample/platformatic.db.json
