+++ packages/client-cli/test/fixtures/sample/plugin.cjs
